pyproject.toml
[project]
name = "amber-content-engine"
version = "0.1.0"
description = "Automated B2B content pipeline for amber — research, topic selection, content generation, review, and publishing"
requires-python = ">=3.11"
dependencies = [
    # LLM — provider-agnostic (both SDKs available, selected via LLM_PROVIDER)
    "openai>=1.30.0",
    "anthropic>=0.42.0",
    "langgraph>=0.4.0",
    "langchain-anthropic>=0.3.0",
    "langsmith>=0.2.0",

    # Database
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg[binary]>=3.2",
    "langgraph-checkpoint-postgres>=0.1.0",

    # API server
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",

    # Scraping
    "firecrawl-py>=1.0",
    "feedparser>=6.0",
    "httpx>=0.28",

    # Integrations
    "gspread>=6.1",
    "google-auth>=2.36",
    "slack-sdk>=3.33",
    "hubspot-api-client>=11.0",

    # Scheduling
    "apscheduler>=3.10",

    # Utilities
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "python-dotenv>=1.0",
    "structlog>=24.4",
    "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
amber-engine = "src.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"
